Move ConsumersResource to V1.
